Fix tests
Jira: MODKBEKBJ-109
create TitlesConverter
